Tables de randomisation multicentre avec taille des blocs aléatoire

Tables de randomisation multicentre avec taille des blocs aléatoire

Usage

listrandph(nbcent = 1, nbtrait = 2, nbcas)

Arguments

nbcent

Nombre de centres

nbtrait

Nombre de classes

nbcas

Nombre total de cas souhaité

Value

1 fichier csv global & un fichier par centre

Examples

listrandph(nbcent = 1, nbtrait = 2, nbcas = 100)